fix: reuse rankings grid card for consistent layout
